Fixed wrong column type on create-products migration

+++ back-end/src/database/migrations/20221130182301-create-products.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(9, 2),
       },
       urlImage: {
         field: 'url_image',
